Move the media picker classes to the mediapicker package
